Update projeject directory
