+++ aws_services/dynamodb_tables/files_per_district_table/coahuila_acereros.json
