+++ aws_services/dynamodb_tables/files_per_district_table/inventory_per_district.json
