+++ aws_services/lambda_functions/data_integration/layers/python.zip
