+++ aws_services/lambda_functions/data_integration/tests/__init__.py
@@ -1,0 +1,1 @@
+from aux_data_integration import *
