+++ aws_services/lambda_functions/data_integration/tests/data/inventory_per_district/coahuila/20230115_acereros_inventory.csv
