+++ aws_services/lambda_functions/data_integration/tests/data/inventory_per_district/coahuila/sabinas_inventory.csv
