+++ aws_services/lambda_functions/data_integration/tests/guide.txt
@@ -1,0 +1,6 @@
+sudo docker system prune -a
+sudo docker ps
+sudo docker build -t diavsimage .
+sudo docker image ls
+sudo docker run -it -v $(pwd):/app diavsimage bash
+pytest -s test_data_integration.py -v --junit-xml=./reports/report.xml --html=./reports/html_report.html --self-contained-html
