+++ aws_services/lambda_functions/data_integration/tests/reports/report.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="1" time="0.199" timestamp="2023-01-31T20:56:00.780896" hostname="664fce088a92"><testcase classname="test_data_integration.TestDataIntegration" name="test_validate_file_extension" time="0.001" /></testsuite></testsuites>
