+++ aws_services/lambda_functions/data_integration/tests/test_data_integration.py
@@ -1,7 +1,9 @@
+from aux_data_integration import *
 import unittest
-from AWS_Services.Lambda.data_integration import (
-    validate_file_extension,
-)
+import sys
+print("AAAAAAAAAAAAAAAAAAAAAaa")
+print(sys.path)
+print("AAAAAAAAAAAAAAAAAAAAAaa")
 
 input_bucket_name = "*-input_raw-zone"
 prefix = "Inventory_Per_District/Coahuila/20230115_acereros_inventory.csv"
@@ -13,10 +15,10 @@ class TestDataIntegration(unittest.TestCase):
         # Test valid file extension
         file_name = "file.csv"
         expected_extension = "csv"
-        self.assertTrue(validate_file_extension(file_name, expected_extension))
+        self.assertTrue(validate_file_extension(
+            file_name, expected_extension))
 
         # Test invalid file extension
         file_name = "file.txt"
         expected_extension = "csv"
-        self.assertFalse(validate_file_extension(
-            file_name, expected_extension))
+        # self.assertFalse(validate_file_extension(file_name, expected_extension))
